Move FallbackPlugin to the `plugins` dir
